requirements.txt
pandas
scikit-learn
pyyaml
fastapi
uvicorn[standard]
joblib
